Revert "refactor: rename model_explainability to ai_safety (#1592)"
This reverts commit 9f72991d7d7707933ae3fa99166aaf338c5f7a09.
