+++ tests/model_explainability/lm_eval/test_lm_eval.py
@@ -1,10 +1,10 @@
 import pytest
 from typing import List
 
-from tests.ai_safety.lm_eval.constants import LLMAAJ_TASK_DATA, CUSTOM_UNITXT_TASK_DATA
-from tests.ai_safety.utils import validate_tai_component_images
+from tests.model_explainability.lm_eval.constants import LLMAAJ_TASK_DATA, CUSTOM_UNITXT_TASK_DATA
+from tests.model_explainability.utils import validate_tai_component_images
 
-from tests.ai_safety.lm_eval.utils import get_lmeval_tasks, validate_lmeval_job_pod_and_logs
+from tests.model_explainability.lm_eval.utils import get_lmeval_tasks, validate_lmeval_job_pod_and_logs
 
 LMEVALJOB_COMPLETE_STATE: str = "Complete"
