+++ tests/model_explainability/trustyai_service/drift/test_drift.py
@@ -2,8 +2,8 @@ from functools import partial
 
 import pytest
 
-from tests.ai_safety.trustyai_service.constants import DRIFT_BASE_DATA_PATH
-from tests.ai_safety.trustyai_service.trustyai_service_utils import (
+from tests.model_explainability.trustyai_service.constants import DRIFT_BASE_DATA_PATH
+from tests.model_explainability.trustyai_service.trustyai_service_utils import (
     send_inferences_and_verify_trustyai_service_registered,
     verify_upload_data_to_trustyai_service,
     verify_trustyai_service_metric_request,
