+++ tests/model_explainability/trustyai_service/service/multi_ns/conftest.py
@@ -12,7 +12,7 @@ from ocp_resources.role import Role
 from ocp_resources.role_binding import RoleBinding
 from utilities.constants import OPENSHIFT_OPERATORS, MARIADB
 from ocp_resources.maria_db import MariaDB
-from tests.ai_safety.trustyai_service.constants import (
+from tests.model_explainability.trustyai_service.constants import (
     TAI_METRICS_CONFIG,
     TAI_DATA_CONFIG,
     TAI_PVC_STORAGE_CONFIG,
@@ -27,11 +27,11 @@ from tests.ai_safety.trustyai_service.constants import (
     GAUSSIAN_CREDIT_MODEL,
     TAI_DB_STORAGE_CONFIG,
 )
-from tests.ai_safety.trustyai_service.trustyai_service_utils import (
+from tests.model_explainability.trustyai_service.trustyai_service_utils import (
     TRUSTYAI_SERVICE_NAME,
     wait_for_isvc_deployment_registered_by_trustyai_service,
 )
-from tests.ai_safety.trustyai_service.utils import (
+from tests.model_explainability.trustyai_service.utils import (
     create_trustyai_service,
     create_isvc_getter_service_account,
     create_isvc_getter_role,
