+++ tests/model_explainability/trustyai_service/service/multi_ns/test_trustyai_service_multi_ns.py
@@ -1,6 +1,6 @@
 import pytest
-from tests.ai_safety.trustyai_service.constants import DRIFT_BASE_DATA_PATH
-from tests.ai_safety.trustyai_service.trustyai_service_utils import (
+from tests.model_explainability.trustyai_service.constants import DRIFT_BASE_DATA_PATH
+from tests.model_explainability.trustyai_service.trustyai_service_utils import (
     send_inferences_and_verify_trustyai_service_registered,
     verify_upload_data_to_trustyai_service,
     TrustyAIServiceMetrics,
