+++ tests/model_explainability/trustyai_service/service/test_trustyai_service.py
@@ -2,20 +2,20 @@ import pytest
 from ocp_resources.namespace import Namespace
 from ocp_resources.trustyai_service import TrustyAIService
 
-from tests.ai_safety.trustyai_service.constants import (
+from tests.model_explainability.trustyai_service.constants import (
     DRIFT_BASE_DATA_PATH,
     TRUSTYAI_DB_MIGRATION_PATCH,
 )
-from tests.ai_safety.trustyai_service.trustyai_service_utils import (
+from tests.model_explainability.trustyai_service.trustyai_service_utils import (
     verify_upload_data_to_trustyai_service,
     TrustyAIServiceMetrics,
     verify_trustyai_service_metric_scheduling_request,
 )
-from tests.ai_safety.trustyai_service.utils import (
+from tests.model_explainability.trustyai_service.utils import (
     validate_trustyai_service_db_conn_failure,
     validate_trustyai_service_images,
 )
-from tests.ai_safety.trustyai_service.service.utils import (
+from tests.model_explainability.trustyai_service.service.utils import (
     wait_for_trustyai_db_migration_complete_log,
     patch_trustyai_service_cr,
 )
