+++ tests/model_explainability/trustyai_service/upgrade/test_trustyai_service_upgrade.py
@@ -1,7 +1,7 @@
 import pytest
 
-from tests.ai_safety.trustyai_service.constants import DRIFT_BASE_DATA_PATH
-from tests.ai_safety.trustyai_service.trustyai_service_utils import (
+from tests.model_explainability.trustyai_service.constants import DRIFT_BASE_DATA_PATH
+from tests.model_explainability.trustyai_service.trustyai_service_utils import (
     send_inferences_and_verify_trustyai_service_registered,
     verify_upload_data_to_trustyai_service,
     verify_trustyai_service_metric_delete_request,
